+++ src/rules/wcag/h67.ts
@@ -20,7 +20,7 @@ class H67 extends Rule {
 			const node = event.target;
 
 			/* only validate images */
-			if (node.tagName !== "img") {
+			if (!node || node.tagName !== "img") {
 				return;
 			}